requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
redis==6.1.0
python-dotenv==1.1.0
pydantic==2.11.4
httpx==0.28.1
